Move unauthenticated views to its own folder
