+++ www/src/routes/Unauthenticated.tsx
@@ -2,10 +2,10 @@ import React, { FC } from 'react';
 import { Route, Switch } from 'react-router-dom';
 
 import { COMMUNITY_ROUTE, MOCK_INTERVIEW_ROUTE, REGISTER_ROUTE, RESUME_REVIEW_ROUTE } from '../util/constants';
-import Community from './Community';
-import Landing from './Landing';
-import MockInterview from './MockInterview';
 import Register from './student/Register';
+import Community from './unauthenticated/Community';
+import Landing from './unauthenticated/Landing';
+import MockInterview from './unauthenticated/MockInterview';
 import ResumeReview from './unauthenticated/ResumeReview';
 
 const UnauthenticatedApp: FC = () => {
